REFACTOR: util to subpackage
